reorganized directory structure
